server/requirements.txt
django
djangorestframework
pillow
radius
psycopg2-binary
drf-yasg # this is swagger for api documnetation